source/back/requirements.txt
asgiref==3.4.1
boto3==1.20.8
botocore==1.23.8
Django==3.2.9
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
jmespath==0.10.0
mysqlclient==2.1.0
Pillow==8.4.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
xlwt==1.3.0
